+++ include/easy_uds/easy_uds.hpp
@@ -7,6 +7,7 @@
 #include "easy_uds/fd.hpp"
 #include "easy_uds/options.hpp"
 #include "easy_uds/request.hpp"
+#include "easy_uds/request_context.hpp"
 #include "easy_uds/response.hpp"
 #include "easy_uds/server.hpp"
 #include "easy_uds/session.hpp"