+++ tests/easy_uds_test/common.hpp
@@ -96,6 +96,7 @@ void test_session_broken_after_timeout();
 void test_idle_session_survives_io_timeout();
 void test_session_move();
 void test_error_model();
+void test_request_context();
 void test_fd_passing();
 void test_reactor_request_timeouts();
 void test_streams();